code signing the wrapper must be done in-place (fixes #13)
Historical note:
Originally I signed wrapper file in resources folder and then copied the files.
Some users reported Asepsis failures when upgrading to 10.9.3:
https://github.com/binaryage/asepsis/issues/13

The kernel is checking mtime of the binary with cs_mtime of code signature, these must match.
Solution is to first copy files to destination folder and execute codesign there.
This should ensure mtime and cs_mtimes are the same.